crcx crc3x: fini methods should reinitialize the lfsr
The LFSR (Linear Feedback Shift Register) holds all of the state necessary for the CRC to be either updated with a new value or finalized.

When the CRC is finalized, the LFSR should be reset to the initialization value so that the same CRC object or context can be reused for repeated, successive CRC computation.

Fixes #64

Signed-off-by: Christopher Friedt <chrisfriedt@gmail.com>